Split module into classes
