+++ lib/bddsm/actual.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module BDDSM
+  class Actual
+    attr_reader :value
+
+    def initialize(value:)
+      @value = value
+    end
+
+    def to(matcher)
+      matcher.match(self)
+    end
+  end
+end
