+++ lib/bddsm/describe.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+# require 'execution'
+
+module BDDSM
+  class Describe
+    def initialize(title, &block)
+      @title = title
+      @block = block
+    end
+
+    def run
+      instance_eval(&@block)
+    end
+
+    def it(&)
+      Execution.new(describe: self, &).run
+    end
+  end
+end
