+++ lib/bddsm/failure.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module BDDSM
+  class Failure
+    attr_reader :error
+
+    def initialize(error, location:)
+      @error = error
+      @location = location
+    end
+
+    def location = @location.join('#')
+  end
+end
